+++ packages/ui/core-components/src/lib/storybook-helpers/WithEvidence.svelte
@@ -21,7 +21,7 @@
 	import '@evidence-dev/tailwind/fonts.css';
 	import '../../../../../../sites/example-project/src/app.css';
 	import { onMount } from 'svelte';
-	import { getThemeStores } from '../themes.js';
+	import { getThemeStores } from '../themes/themes.js';
 
 	const { syncDataThemeAttribute } = getThemeStores();
 	onMount(() => syncDataThemeAttribute(document.querySelector('html')));
