+++ packages/ui/core-components/src/lib/unsorted/viz/box/Box.svelte
@@ -10,7 +10,7 @@
 
 	// import { prepareBoxplotData } from 'echarts/extension/dataTool';
 	import { formatValue } from '@evidence-dev/component-utilities/formatting';
-	import { getThemeStores } from '../../../themes.js';
+	import { getThemeStores } from '../../../themes/themes.js';
 
 	const { resolveColor } = getThemeStores();
