+++ packages/ui/core-components/src/lib/unsorted/viz/core/ECharts.svelte
@@ -12,7 +12,7 @@
 	import ChartLoading from '../../ui/ChartLoading.svelte';
 	import { flush } from 'svelte/internal';
 	import { createEventDispatcher } from 'svelte';
-	import { getThemeStores } from '../../../themes.js';
+	import { getThemeStores } from '../../../themes/themes.js';
 
 	const { activeAppearance, resolveColorsObject } = getThemeStores();
