+++ packages/ui/core-components/src/lib/unsorted/viz/map/PointMap.svelte
@@ -6,7 +6,7 @@
 	import Points from './components/Points.svelte';
 	import BaseMap from './_BaseMap.svelte';
 	import { Query } from '@evidence-dev/sdk/usql';
-	import { getThemeStores } from '../../../themes.js';
+	import { getThemeStores } from '../../../themes/themes.js';
 
 	const { resolveColorPalette } = getThemeStores();
